Split tests in several test files
